backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0

